feature: move search logic to new column
