+++ frontend/src/components/HomePage.module.scss
@@ -15,8 +15,8 @@
   justify-content: center;
   flex-direction: row;
   align-items: center;
-  width: 75dvw;
-  height: 80dvh;
+  width: 90dvw;
+  height: 90dvh;
   overflow: hidden;
   border-radius: 2rem;
 }
