+++ frontend/src/components/HomePage.tsx
@@ -5,6 +5,7 @@ import styles from './HomePage.module.scss';
 import ChatWindow from './chatWindow/ChatWindow';
 import FriendList from './FriendList/FriendList';
 import { FriendProvider } from '../contexts/FriendContext';
+import CommunityPage from './communitypage/CommunityPage';
 
 const HomePage = () => {
   const { loggedIn } = useContext(AuthContext);
@@ -14,6 +15,7 @@ const HomePage = () => {
       {loggedIn ? (
         <div className={styles['main-chat--wrapper']}>
           <FriendProvider>
+            <CommunityPage />
             <FriendList />
             <ChatWindow />
           </FriendProvider>
