+++ frontend/src/components/communitypage/CommunityPage.module.scss
@@ -1,4 +1,4 @@
-.friendlist--wrapper {
+.community-page--wrapper {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -8,29 +8,23 @@
   box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
   flex: 2;
   min-height: 100%;
-  overflow: scroll; /* Prevent content from spilling out */
+  overflow: scroll;
+  overflow-x: hidden;
 }
 
-.friendlist--title {
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: #333;
-  margin-bottom: 1rem;
-}
-
-.friendlist--container {
+.community-page--container {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
   gap: 1rem;
   width: 100%;
-  height: 100%; /* Let it take up all available height */
-  max-height: 400px; /* Or any desired height for scrollable area */
-  overflow-y: auto; /* Enable vertical scrolling */
-  padding: 1rem; /* Optional: add some padding inside */
-  box-sizing: border-box; /* Ensure padding doesn't affect dimensions */
+  height: 100%;
+  max-height: 400px;
+  overflow-y: auto;
+  padding: 1rem;
+  box-sizing: border-box;
 }
 
-.friend-card {
+.community-page-card {
   cursor: pointer;
   background-color: #ffffff;
   padding: 1rem;
